Create GUI.py, ModeMenu.py, and update MainMenu.py
GUI is a class that deals with a menu's visual elements.
It keeps a dict of the elements and displays them in 'view'.
The event handling system in GUI checks if a Button element
is clicked and calls/returns an action.

MainMenu is updated to inherit GUI.
ModeMenu is an instance of GUI for the mode (classic/survival)
selection menu.